Refactor user registration logic: Separate controller and routes
