+++ controllers/userController.js
@@ -1,0 +1,39 @@
+const bcrypt = require("bcryptjs");
+const User = require("../models/User");
+
+const registerUser = async (req, res) => {
+  try {
+    const { name, email, password, role, branch_id } = req.body;
+
+    const existingUser = await User.findOne({ where: { email } });
+    if (existingUser) {
+      return res.status(400).json({ error: "Email already registered" });
+    }
+
+    const validRoles = ["admin_central", "branch_manager", "branch_employee"];
+    if (!validRoles.includes(role)) {
+      return res.status(400).json({ error: "Invalid role" });
+    }
+
+    if (role !== "admin_central" && !branch_id) {
+      return res.status(400).json({ error: "Manager and employee must have branch_id" });
+    }
+
+    const hashedPassword = await bcrypt.hash(password, 10);
+
+    const newUser = await User.create({
+      name,
+      email,
+      password: hashedPassword,
+      role,
+      branch_id: role === "admin_central" ? null : branch_id,
+    });
+
+    res.status(201).json({ message: "User created", user: newUser });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ error: "Error creating user" });
+  }
+};
+
+module.exports = { registerUser };
